Fix `Paths` to ignore platform conventions when being told so
I mistakenly assumed `ProjectDirs::from_path` would ignore the platform-dependent
roots (e.g. XDG_* on Linux), but it instead overrides any paths _below_ that.

What we actually want is to be able to store all configuration and state under
a user-defined root directory, which is useful for testing or otherwise
constrained deployments. So this patch fixes that by precomputing things
either in a platform-dependent (`new()`) or user-defined (`from_root`) way.