Replace log with tracing.
This introduces tracing into librad to replace log.
We import it qualified so that it makes it easy to use any of the
functionality from the lib with updating the imports.

tracing allows us to introduce spans at a top-level so that common
information can percolate down.

The extra benefit to tracing is that we can pass in key, value pairs to
trace statements for structured logging purposes.