Check if a fetch triggered by gossip actually applied the value
Otherwise, pretend we've already seen it -- which just means to stop propagating
the gossip here. We know at this point that the provider either lied, or the ref
was already gone before we could fetch it. In either case, and since there was
no error, noone else should be able to get the data from the provider either.
And since we don't have it, neither from us.