Introduce a local transport for interacting with the storage backend
This simply connects to stdio of forked `git-upload-pack` / `git-receive-pack`
processes, which are started with a configuration suitable so the caller can
only access the current namespace. Remote tracking branches are hidden for
`receive-pack`, such that pushing to them is prohibited. The transport is set up
with a signer (ie. secret key), such that `Storage::update_ref` can be called
transparently after a successful push.

In short, interacting with a logical repo requires simply registering the
transport, and setting up a remote with a URL of the form `rad://<id-part-of-URN>`.

Note that the transport is set up in a way so it can be reused for a git remote
helper, allowing CLI access to the storage backend. The helper code will be
submitted in a subsequent patch.

We can now also remove the `Deref<Target = git2::Repository>` from `Storage`,
prohibiting access to the raw git handle.