Type witness Reference
Adding phantom data to Reference we can distinguish between one that
points to a single reference, one that points to multiple references, or
one that can point to any type of reference.

This means the API will ensure that we call 'find' on a single
reference, 'references' on a Reference that points to multiples
references.

The Any witness is used for Refspecs since it needs to be able to store
both types in a Vec when using fetch_heads.

The final change is that Storage surfaces the functions for 'find' and
'references' so that the caller can build these Reference types and use
them safely.