Permeate the Signer
As we introduce the use of Signer::sign it begins to tug on all the
higher levels of the stack. This change replaces the use of SecretKey
with the Signer trait and call to Signer::sign. This meant that a lot of
the bounds had to change.
As well as this the Signer has an associated Error type which ends up as
a dyn Error in the enums. This means we also have to add bounds for the
associated error.

One thing I had to add was the trait AsPKCS8 to allow us to get the
PKCS8 bytes ad-hoc for the network calls. Not sure if this was the right
call or not yet.

The final thing to address is that the local transport settings have a
static ref. This holds a SecretKey and I'm not sure how make this
generic.