Refactor traits
Refactored WithSigner to be a blanket implementation for a lot of the
common traits we need the signer to be across the stack.
As well as this we introduce a blanket trait 'SignError' for the
associated error of the Signer.
Also fixed up the use of PeerId::from_signer so as not to use
`Into<PeerId>`s everywhere.