Introduce WithSinger for local/transport.rs
This change may be controversial. The idea of a global static variable,
i.e. `SETTINGS`, was removed because it was proving to be too difficult
to massage using `Box` et al. Instead we removed it and pass around the
Settings via function params.

The registering was tentatively tested by calling twice it
propagation_basic without any error occurring. Whether this was good
enough, I'm not sure...