Box dyn magic for local transport
The first step was to remove Clone from the blanket Signer trait. This
was getting in the way because Clone requires Sized, preventing us from
creating a trait-object.

Then, by creating a newtype over a Box<dyn Signer> we can implement
Clone and Signer allowing us to use it as the parameter for Storage.
Note, however, that we also need to give it a concrete Error. So we give
it a Box<dyn Error>.

The final piece is introducing SomeSigner which knows how to implement a
keystore::sign::Signer with a BoxedSignError for any S that already does so.