Query for URN
The previous attempts would not give the desired result due to having to specify
the provider PeerId ("origin"). It turns out we can make the origin optional,
where the `ask` answer is positive iff the peer has the requested value in its
own view of the repo. However, if we really only know the URN, we don't have to
wait for exactly this answer -- any `Have` with a matching URN can be tried. For
this we need to know what the gossip payload type is instantiated to, so we move
this specific query to the top-level `PeerApi`.

Also, we need to propagate the address info (ie. `PeerInfo`) for `Have` messages.