Determine remote identity when TLS session is resumed
When a TLS session is resumed (which is enabled by default), rustls doesn't
remember the presented certificate, and doesn't provide a way to attach
additional session persistence information. However, in the case of incoming
connections, we can still fall back to the SNI (relying on the cert verifier to
require SNI). Outgoing connections know the SNI already, so they can use that
by the same reasoning.