Factor out the namespacing
The aim of this change was to factor out the common structure of a
namespaced reference and a top-level reference.

This required some messing around with SomeReference so that we could
talk about SomeNamespace. This was necessary to massage the type
checking.

The real API is then exposed through a FlatRef, which has no namespace,
or a NamespaceRef which is parameterised over the Namespace hash.