Polymorphic refspec constructor
In our quest for MORE POLYMORPHISM, we move the refspec constructor of
two references into the polymorphic methods of Reference. We also allow
the remotes to change in a Refspec.
This means that we can pass in a Reference with a different Namespace
and Remote, and get back the Refspec for it.
The bonus is that the equality of the Cardinality of the two References is
enforced by the ReferenceInfo trait.