String type for Canonical JSON
To ensure equivalence, we need to store the string in Unicode
Normalization Form C (which OLPC Canonical JSON serialises to). To make
`serde_json` actually parse JSON documents in canonical form (which is
not a proper subset of RFC 7159 JSON), we need to parse to a
`serde_bytes::ByteBuf`, and perform the normalisation on the unchecked
utf8 representation.