Use a PeerId Map
Instead of a global lock, we have a map from PeerId to Settings.
The PeerId can be easily calculated from Settings when registering,
since Settings contains a signer.
The annoying change is that we have to add a PeerId to LocalUrl, where the
PeerId is the username of the URL. This means we have to pass around
PeerIds to LocalUrl transformations.