Use a HashMap for shared Settings ref
The local transport was using a global mutex on `Option<Settings>`. In
concurrent runs tests would end up overwriting each other's version of
this mutex and would end up in inconsistent states.

To solve this we need to separate test case's runs. Initially, we
attempted to use another lock `Mutex<()>` and have a bracketed runner,
which would lock, fill the `Option<Settings>`, run the action, and clean
up & release the lock. This still ended up in race conditions.

This solution uses a HashMap from PeerId to Settings. This works since
we expect individual runs to have separate PeerIds. The consequence of
this is that we need a way to look up the Settings when running the git
action. We encode this in the LocalUrl, so we add the PeerId as the
username of the Url that's passed.