Add Include
An Include is a way for users to easily manage creating an include file
for their working copies. Currently it's simply a list of Remotes, which
point to peers. This will generate the include file at a provided
directory. The user can then simply link the include into their working
copy config.