Fix advertised remote refs
Two fixes to the advertisement of refs from the monorepo were needed.
The first was the glob that was passed for finding visible remotes
didn't include "refs/" and only needs "*" instead of "**".
The second was to remove the '^' operator in the uploadpack.hiderefs
argument. This was preventing the refs from being advertised as
expected.