Savely overwrite include files
Until now git2s `Config` was used to handle the actual file on disk for
includes. While an intuitive approach it makes it hard to reason about
the state of the actual file, as reliably replacing the entire content,
if for example the tracking graph has changed drastically, would require
building a diff and then invoke `remove` on per remote basis.

As a drop-in replacement `Include` now offers a `save` method which --
while consuming the struct -- writes the content to temporary file and
then atomically replaces the old version.