Testing 3 peers cloning
We add a test that does the following:
* Peer 1 creates a project
* Peer 1 makes a commit to their default branch
* Peer 2 clones from Peer 1
* Peer 3 clones from Peer 2

The expectation is that Peer 2 has Peer 1's heads/*, which it does.
It is also expected that Peer 3 has Peer 1's heads/*, which it doesn't.

We were only checking at the top-level for heads, i.e.
refs/namespaces/<namespace>/refs/heads/*, but we weren't checking for
remote entries, i.e.
refs/namespaces/<namespace>/refs/remotes/<peer>/heads/*.