Audit pass by reference
We went through the functions that were taking PeerId as a parameter and
decided whether a reference was needed or if it should be owned. This
was decided by whether only a reference of PeerId was required in the
function body, if so then it is passed by reference. If the PeerId is
needed to owned or copied, then an owned PeerId is passed down.