Existence check and adoption
After the fetch we ensure that we did indeed fetch rad/id for each
certifier of the urn under remote_peer. If the certifier does now exist
we adopt the rad/id of the remote_peer as our own (if it hasn't been
cloned already).

We don't check rad/ids/* due to signing by OwnedKey not creating a
rad/ids entry and this code generally being broken with new code taking
its place sooooooon.