Fix certifier linking
We were trying to parse a Hash from a fully formed reference. Instead we
should have been using the urn_from_ref helper to get the RadUrn.

We were missing the remote for the rad_certifier.

We had the symbolic_ref the wrong way around, i.e. the ref that gets
passed as a parameter is the one that gets created.