Local transport: side-channel for blocking on local transport results (#377)
`libgit2` may drop all references to the local transport before the
child process exited. We must, however, `wait(2)` on the children on a
(non-cancellable) thread, which leads to the dilemma of either risking
to block the parent process (by joining on this thread), or exiting it
prematurely.

We thus establish a side-band channel which allows non-longrunning
processes to fully await completion of their local transport operations
(e.g. in a `Drop` finaliser).