Add a check that the certifier is fetched (#374)
* Add a check that the certifier is fetched

This currently shows that it isn't working as understood but instead
panics at the call to some_metadata.

* Construct RadUrn

The previous code was trying to parse a Hash as a RadUrn. Instead, we
want to parse the Hash and then construct the RadUrn with the result.

* Fetch for each certifier

* Relax uploadpack.hiderefs

In p2p::server we were set up in the upload-pack call with an argument
of the form `uploadpack.hiderefs=!refs/namespaces/{}/rad/id`. This is
missing a `refs/` before `rad/id`. But we can also just relax it so that
it's scoped to just `refs/namespaces`.

* Remove call to fetch_repo

* Add FIXME

* Clean up Clone and format

* Existence check and adoption

After the fetch we ensure that we did indeed fetch rad/id for each
certifier of the urn under remote_peer. If the certifier does now exist
we adopt the rad/id of the remote_peer as our own (if it hasn't been
cloned already).

We don't check rad/ids/* due to signing by OwnedKey not creating a
rad/ids entry and this code generally being broken with new code taking
its place sooooooon.

* Fix certifier linking

We were trying to parse a Hash from a fully formed reference. Instead we
should have been using the urn_from_ref helper to get the RadUrn.

We were missing the remote for the rad_certifier.

We had the symbolic_ref the wrong way around, i.e. the ref that gets
passed as a parameter is the one that gets created.

* Adopt certifier as direct reference

* Adopt certifier as ours

If we do not have `refs/namespaces/<certifier>/rad/id` we adopt
`refs/namespaces/<certifier>/refs/remotes/<remote peer>/rad/id` as our
own. Note that this should actually run verification, but wtf is a
resolver anyway.

If/when `refs/namespaces/<certifier>/rad/id` exists we create a symref
for `refs/namespaces/<current project>/rad/ids/<certifier> ->
refs/namespaces/<certifier>/rad/id`.

* Fix NamespacedRef parameters