Scope local PeerId cases
Instead of passing the local_peer_id to urn_context, we handle the cases
scoped to put and ask functions.
In the case of put, we should return Stale since the local peer should
know about its own changes.
In the case of ask, we check that the origin is local PeerId, and if so
we say we point the Urn to our top-level (as opposed to Originates).