Peer LocalStorage fixes (#401)
* Qualify paths in both cases of urn_context

* Add logging

* Pass down provider

* Revert logging changes

* Check if the path is empty

* Run formatter

* Propagate the applied value

* Deduplication of origin resolution

* Urn still needs to point to correct PeerId on remote side

If the origin was None, it means that we should point to the original
Urn when fetching. We then set the origin to the PeerId of who we fetch
from, either the provider or the origin.

* Do not touch refs/rad paths

* Scope local PeerId cases

Instead of passing the local_peer_id to urn_context, we handle the cases
scoped to put and ask functions.
In the case of put, we should return Stale since the local peer should
know about its own changes.
In the case of ask, we check that the origin is local PeerId, and if so
we say we point the Urn to our top-level (as opposed to Originates).

* Lock up and lock down

* Revert "Lock up and lock down"

This reverts commit 4d88f31bc23a790930fbfc8dd89666af1fdc1745.

* Revert "Scope local PeerId cases"

This reverts commit 16d1829ac7c4376249eec8f9a2a0fd4e5aec380f.

* Formatting docs

* Backport changes

Using
https://github.com/radicle-dev/radicle-link/commit/b302017c5d8351898b96e9f439a632ef11a489b2#diff-b650ce3d62e812ad5579e43d0805c510f7324e38f24762a9a81f5466b760993bR468
as inspiration, we use RefLike and Qualified to logically qualify the
RadUrn paths.

* Backport RefLike strip_prefix

* Backport tests and add fix

In
https://github.com/radicle-dev/radicle-link/commit/b302017c5d8351898b96e9f439a632ef11a489b2#diff-b650ce3d62e812ad5579e43d0805c510f7324e38f24762a9a81f5466b760993bR705
there is a suite of tests that are useful for confirming that
urn_context works as expected. We grab those and alter them to work
RadUrn's. This lead to a fix in the current implementation.

* Log for which urn and peer we're getting signed_refs for

* Add sign-refs tool

Adding a tool so that we can easily verify if we're getting the right
calculated signature for signed refs. It takes the refs as JSON input,
signs it with the key in your config, and presents the Signed output.

* Allow the caller pass the key via stdin

* Revert "Allow the caller pass the key via stdin"

This reverts commit 77d1f8866b1b24b676999a36d2bc3d2403699dcf.

* Revert "Add sign-refs tool"

This reverts commit e8ec032a4bb6e4c98c7d6668bfdf104334c82a13.

* Log Signed

* Use eprintln for rad remote tracing

* Revert "Use eprintln for rad remote tracing"

This reverts commit 02c31a31673c3bc8fc269917b2c614440ddd6cf6.

* Revert "Log Signed"

This reverts commit c6d2099d4a164271435b975707e595bc41636132.

* Remove logging