Add Info::Applied
When having tracked a peer we don't immediately have their changes. We
can know that we do have the peer's changes when we have fetched their
changes or if we have applied them from a gossip message.

We need a way to surface that a result was successfully applied. In this
patch we do this by adding it to the `Info` enum and emit the protocol
message to the subscribers.