Use BTreeSet instead of HashSet
Using a HashSet the iter method will return items in an arbitrary order.
Since this is what serde uses under the hood to turn HashSet into an
array, it destabilises the signing and verification of refs.

Using a BTreeSet gives us a guaranteed order for the array, meaning
consistent de/serialization across peers.