Include fixes
If no fetch refspec was provided we would incorrectly use our own PeerId
in the refspec. This is essentially useless because we shouldn't have
our own PeerId under our remotes.

Instead, we change the interface so that remotes are added by providing
the PeerId and handle, then building the Remote with the fetch refspec
internally.