local transport: access via "bracket"
Due to post-service hooks and the possibility to trigger GC after a
push, the local transport is inherently racy, because `libgit2` drops
the handle too soon (as soon as it read the expected output). When the
network latency is low, this may e.g. cause announcements to be received
too early if they are made from a callback. There is considerable
ceremony involved to prevent that.

We give up on interruptible wait(2), and instead enforce the use of a
"safe" API through the `git::types::Remote` wrapper.

Collaterals:

* Use trait object for Storage signer

  Turns out to be the less painful workaround for the impossible arrow
  `&Storage<S: Signer> -> &Storage<BoxedSigner>`

* Refine push and fetch spec types

  A push spec MUST resolve to a single ref (ie. it can not contain any
  patterns). libgit2 doesn't seem to support expressions (e.g. `HEAD~2`),
  and neither do we. Mirroring git push --all, we do, however, allow
  resolving matching refs from the local repo.

  It is also rather confusing to talk about a "local" and "remote" end
  of a Refspec, as that depends on whether it is used as a fetch or push
  spec -- instead, use git terminology (src, dst), and force "named
  args" throughout.

* Demote namespacedness of the Reference type

  It turns out that there are quote surprising interactions with impl
  resolution and inference, due to fundamental types and PhantomData not
  being a local type. As we don't have a use case (that I'm aware of)
  which would require distinguishing namespaced and non-namespaced refs
  statically, just storing an Option (again) seems way more, errr,
  errrgonomic.