net: reject duplicate connections
The idea to eject previous connections from a remote peer is racy, as
described in #431. Instead, we reject new incoming connections until the
old one is either closed gracefully, or times out.

This also entails to not automatically terminate connections when a peer
is ejected from the active set of gossip membership.

Closes #431
Signed-off-by: Kim Altintop <kim@monadic.xyz>