Replaces internal representation of references to avoid using pathbuf
The internal representation of RefLike and Friends used the convenient
PathBuf methods to join, split and traverse the components in a
path-like manner. Unfortunutally it [does not work on
windows](https://github.com/radicle-dev/radicle-link/issues/467) as it
uses a different path separator.

This commit changes the internal representation from using pathbufs to
use string internally. It makes the conversion to string more
convenient, as it will always be valid utf-8, but it makes manipulation
a bit harder, as we have to remember to add the trailing `/` on strip
prefixes and avoid them on join.

More commits fixing the compilation to follow.

Fixes https://github.com/radicle-dev/radicle-link/issues/467

Signed-off-by: Bruno Tavares <connect+github@bltavares.com>
Signed-off-by: Kim Altintop <kim@monadic.xyz>