Fix helper lookup test on windows
The test helper was stripping the suffix to get the full path of the
directory, which does not work on windows as binaries add the `.exe` on
the end.

This commit uses the PathBuf methods to navigate to the path instead of
hardcoded strings to make it cross-platform. Tested over WSL as well.

Signed-off-by: Bruno Tavares <connect+github@bltavares.com>