Avoid deleting the temporary folder before finishing tests
The library used to create temporary folders has a cleanup behaviour, so
that the folder is removed when the reference is dropped.

Somehow, on Linux, it like-ly didn't trigger as we had a `mkdir -p`-like
call right after it, but on Windows, it was deleting the folder before
the tests finished.

Instead of moving the tempdir reference into the call to then drop it
internally, this commit uses them as references so we keep the variables
living longer, until the end of the test.

Signed-off-by: Bruno Tavares <connect+github@bltavares.com>