Fix strip_suffix calls due to internal representation changes
Previously, when the internal representation of RefLike was a PathBuf,
we would be able to strip prefix with or without the trailing `/`.

It would then remove the component accordingly, including any `/` that
would be hanging.

Now that the implementation uses a String, we need to strip_prefix
always using a `/`. This was not easy to find tho, there were tests
failing but it was not clear why until the sudden realisation.

We are covered by tests - but the error was not clear.

Signed-off-by: Bruno Tavares <connect+github@bltavares.com>