Fix clock tests on windows
std::time on Rust is not OS-agnostic, and it uses an internal OS C
struct for SystemTime.

On windows, this structure only handles i64 values, not the usual u64
from Unix, and the test triggred an overflow when running prop tests
with a too large number.

This commit changes the range of test values to something that works on
both OS - but it's still a large number of years

Signed-off-by: Bruno Tavares <connect+github@bltavares.com>