Use trustee instead of remote
When we are fetching, we don't need the remote for any of the
verification or detecting of forks, instead we use our local version of
the identity.

As well as this, we depend on one of the delegates or tracked peers as
the new view of the world if delegates were updated.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>