Fetch and verify what we know
Instead of using a "trustee" as the source of truth for the new
delegates we just fetch the set that we know and track any new peers. As
the TODO at the bottom of replicate outlines, we will eventually
converge and fetching of new delegates could be done as some kind of
callback mechanism.

For now, we just fetch the delegates and tracked peers that we know,
track new ones. For adopting the most up-to-date identity we first check
if we are a delegate ourselves. If we are not then we adopt the latest,
otherwise, we return a data type that indicates that user intervention
must happen for updating the document.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>