net: cache mDNS responses
The mDNS discovery will discover they same set of peers every
`query_interval` (if they are still responding). To not interfere with
the membership protocol, it should not yield repeatedly discovered peers
(at least for a while): otherwise, the connectivity of the local peer
would converge to the fastest responders on the local network, and
nothing else.

Signed-off-by: Kim Altintop <kim@monadic.xyz>