Pass fetch limit at runtime
We pass controlling of fetch limits up to the runtime. We do this by
introducing a new struct `Limit` which controls an absolute limit, for
when we PeekAll, and relative for when we know the number of refs to
expect.

This control is needed for `replicate` and is fed down from
`PeerConfig`, through `PeerStorage`, and `PeerApi`.

We provide a Default where the relative number is 500KiB and the
relative is 1KiB.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>