Replication Improvements
The reason for this set of changes stemmed from replication missing some
references that were expected to be present. This lead us to exploring a
different way of organising the replication.

The change introduces:
* Separate replication types: clone and fetch
* A clone is a fresh fetch from the remote where we fetch all the
necessary data and prune after
* A fetch is fetching the remotes we're already aware of and updating
the identity based on that information.
* Our verification now ensures that all identities are verified, are not
forks, and then we adopt the latest tip from the delegates.
* We add a configurable limit to fetching so that we can tune how much
data we're allowing to be fetched from a remote.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>