Introduce radicle-copy
radicle-copy is the API for creating projects in the monorepo _as well_
as a working copy to mirror it. It allows the creation of a fresh new
project or a project based off of an existing git repository.

A part of working copy management is also generating and maintaining
include files that persist the tracked remotes.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>