Use fetch instead of download
We have to use fetch instead of download due to a bug that is resulting
in a bad negotiation between two clients where an extra `want` is
issued after a `have`. Some more details are here
https://github.com/libgit2/libgit2/issues/5799.

This should only be temporary because this means we fetch the refs
advertisement multiple times, redundantly.

We need to have the update tips logic be part of the callback,
otherwise the returned updated tips will always be empty, due to
`Remote::update_tips` being a no-op after `Remote::fetch`.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>