instrumentation: stream error span + URN display
Fixes a few instances of URNs going through Debug instead of Display for
logging. This is still not ideal, because composite values are often
better logged through Debug, but then the URN field is somewhat less
useful again. Need to investigate with upstream why the Value trait is
sealed.

Also improves the span reporting when a stream error occurs.

Signed-off-by: Kim Altintop <kim@monadic.xyz>