git, net: Add a "capability" trait to mark pooled storage
It turns out we will need access to a `git::Storage` from some stream
handlers, specifically in pooled form due to `Send`/`Sync` requirements.
As we need to pull that out from a larger structure (respectively
wrappers), a trait is in order. Since we need to thread the constraint
through the various IO function, we defined a composite trait to make
constraint signatures shorter.

Signed-off-by: Kim Altintop <kim@monadic.xyz>