git-ext, storage: Add a way to quickly create a commit
Mainly in integration tests, we often want to create some git history.
Currently, the only way to do that is to go through a working copy,
which exercises that flow, but is often inconvenient (and slow).

Since we want to avoid handing out a raw git handle to the underlying
"monorepo" storage, we instead expose a public means to create a commit
in a reasonably safe and almost ergonomic way.

Signed-off-by: Kim Altintop <kim@monadic.xyz>