sync: introduction of fetch and announcment
Introduces the announce and fetch modules that are ported over from
proxy::coco (announcement and sync, respectively).

The announce module has logic for saving, loading, and emitting new
tips based off of a URN's Refs.

The fetch module has logic for attempting synchronise the storage based
off of its current Refs for a URN (or URNs).

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>