net: Mutual fetch (grafting pt. 1)
As the first step towards the grafting protocol (aka gossip-independent
bulk sync), a peer now may initiate a reverse replicate (rere) whenever
a git fetch comes in. Whether a rere is initiated depends on an
intersection of locally tracked and remotely advertised `PeerId`s.

To prevent this procedure from looping indefinitely, the initiator now
includes a random nonce in the fetch request header. Only if the nonce
is present and determined to be reasonably fresh, a rere is attempted
(which, in turn, does **not** set any nonce). This nonce may come in
handy in other ways, too.

Note that the nonce is generated by default, but it is possible to
explicitly avoid a rere by disabling it.

Signed-off-by: Kim Altintop <kim@monadic.xyz>