Properly fetch tag refs
We now include tags properly in the list of refs to be fetched from a
peer.

When we want to fetch monorepo references from a remote
`refspec::sigrefs` generates the list of fetch specs to use from the
remote’s refs and the refs that the local peer is interested in. Before,
the reference category (i.e. `tags`, `heads`, `rad`, etc.) was ignored
and replaced with `heads` when creating the fetchspec. For example, if
the remote advertised `refs/tags/foo` and the local peer was interested
we would add `refs/heads/foo` to the fetchspec.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>