git: Prevent concurrent fetch of the same URN
It is in principle safe to fetch multiple packfiles concurrently, even
if their contents end up being identical. However, updating the refs
after a fetch is not an atomic operation, and even though libgit2
tries to detect concurrent modifications, operations which rely on
read-after-write semantics will not function properly.

To fix this, we effectively serialise replicate of the same logical
repository (ie. URN). There are several tradeoffs, as documented in the
code.

As for continuous brooming, the git::fetch module is broken up into
more submodules for reduced strain on scrolling devices.

Signed-off-by: Kim Altintop <kim@monadic.xyz>