storage: check for tags in has_commit
The use of `find_commit` meant that we were ignoring tag objects. This
patch uses `find_object` instead to check if we have `Commit` or `Tag`
and then confirms if the SHA is in the given reference.

We extend the gossip test to see if the gossip of the tag gets propagated
back to `peer1`.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>