daemon: migrating the radicle-upstream/coco code
We're moving the code that was living in radicle-upstream/coco to start
as the foundation of a daemon.

The one caveat is that coco/source.rs will not be present as its
migration path is towards radicle-surf.

Everything else is a verbatim move of the coco files where the only
changes were for imports. These were mostly using `librad::git_ext` as
opposed to `radicle_git_ext` and the removal of `radicle_surf::git2`.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>