net: Interrogation API (grafting pt. 2)
In the second installment towards the grafting protocol, we introduce an
API and associated stream type to interrogate a known remote peer about
certain data. Of relevance for grafting is strictly speaking only the
RPC to ask for the complete list of URNs.

We use the Xor filter probabilistic datastructure for that, in order to
gain compact and bounded responses (see commentary inline). The work to
compute the set intersection will be on the requester, while the
responder recomputes the filter only when necessary (morally; we can't
currently intercept refs updates in the relevant places, so recompute
after an expiry timeout).

Signed-off-by: Kim Altintop <kim@monadic.xyz>