net: advance membership on connection errors
When switching from long-lived bidi- to ad-hoc uni-streams for gossip,
it was not properly considered that the partial view will not repair
itself in mostly-idle networks -- when there's nothing to send, and
nothing received, we need a cue from the connection driver and dispatch
that to the membership state machine. Note that this relies on PTO
expiration respectively delivery of CONNECTION_CLOSE frames from the
remote, and thus doesn't address any weird states due to "unclean"
reconnects within the PTO period.

Related-to: #575
Signed-off-by: Kim Altintop <kim@monadic.xyz>