smoke tests: stress test the fetch limit
The fetch limit for peeking is set to 5KB. This doesn't take into
account that the refspecs could ask for multiple remotes -- say if
there was multiple maintainers. This test sets up multiple remotes
under a mainters project, then the final peer tries to replicate it
and fails due to the limit being reached.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>