testing: allow pointing integration tests to any bootstrap node(s)
This is mainly intended to be able to bootstrap against a large-ish
compose network, and will refuse to join any other network than
"localtestnet". Observe that some tests indeed fail when doing so.

The testnet module is also reworked to use the notorious RAII pattern,
instead of a safe bracket. This makes tests less robust, but is closer
to reality. Observe that it indeed surfaces a few cancellation issues.

Lastly, the example is removed from librad. This had been forgotten, as
it was an intermediate step before the e2e "ephemeral-peer" was
introduced.

Signed-off-by: Kim Altintop <kim@monadic.xyz>