git: use dedicated runtime for p2p transport
This removes nesting of executors, which should make things less
susceptible to hangs because the right incantation of
spawn/spawn_blocking was not adhered to. At least, it improves
reasoning, and allows for some instrumentation.

It may or may not fix the 12 monkeys issue. For that, it may be
advisable to manage _all_ runtime instantiations from librad, instead of
relying on ambient executors.

Signed-off-by: Kim Altintop <kim@monadic.xyz>