net: use a bounded vec for propagated address info
This is a long-standing open issue: we need to make sure that collection
types are always bounded on the wire.

Note that, while morally a set, set semantics were not captured
precisely:

1. For establishing a connection, we take the union of the seen and
   advertised addresses, with a bias towards the former.
2. For maintaining the seen addresses, we bias towards the more recent
   information.

1. needs to be computed at runtime anyways, while 2. is much simpler to
maintain using a (small) Vec instead of a specialised insertion-ordered
set type. We thus spare ourselves the nuisance of having to thread
through an Ord or Hash constraint on the datatype level.

Signed-off-by: Kim Altintop <kim@monadic.xyz>